Refactor: Moved SimulationLibraries
