+++ diffsims/libraries/diffraction_library.py
@@ -17,21 +17,14 @@
 # along with diffsims.  If not, see <http://www.gnu.org/licenses/>.
 
 import pickle
-from typing import NamedTuple, Sequence, Set
 
 import numpy as np
 
-from orix.quaternion import Rotation
-from orix.crystal_map import Phase
-
-from diffsims.sims.diffraction_simulation import DiffractionSimulation
 from diffsims.generators.diffraction_generator import DiffractionGenerator
 
 __all__ = [
     "DiffractionLibrary",
     "load_DiffractionLibrary",
-    "SimulationLibrary",
-    "SimulationLibraries",
 ]
 
 
@@ -100,59 +93,6 @@ def _get_library_entry_from_angles(library, phase, angles):
     raise ValueError(
         "It appears that no library entry lies with 1e-2 of the target angle"
     )
-
-
-class SimulationLibrary(NamedTuple):
-    phase: Phase
-    rotations: Rotation
-    diffraction_generator: DiffractionGenerator
-    simulations: Sequence[DiffractionSimulation]
-    str_rotations: Sequence[str] = None
-
-    def __repr__(self):
-        return (
-            f"DiffractionPhaseLibrary(phase={self.phase.name},"
-            f" No. Rotations={self.__len__()})"
-        )
-
-    def __post_init__(self):
-        if len(self.rotations) != len(self.simulations):
-            raise ValueError("Number of rotations and simulations must be the same")
-        if self.str_rotations is not None and len(self.rotations) != len(
-            self.str_rotations
-        ):
-            raise ValueError("Number of rotations and str_rotations must be the same")
-
-    def __len__(self):
-        return len(self.rotations)
-
-    def __getitem__(self, item):
-        if isinstance(item, str):
-            item = self.str_rotations.index(item)
-        return SimulationLibrary(
-            self.phase, self.rotations[item], self.simulations[item]
-        )
-
-    def get_library_entry(
-        self, rotation: Rotation, angle_cutoff: float = 1e-2
-    ) -> "DiffractionPhaseLibrary":
-        angles = self.rotations.angle_with(rotation)
-        is_in_range = np.sum(np.abs(angles), axis=1) < angle_cutoff
-        return self[is_in_range]
-
-
-class SimulationLibraries(dict):
-    """
-    A dictionary containing all the structures and their associated rotations
-    """
-
-    def __init__(self, libraries: Sequence[SimulationLibrary]):
-        super().__init__()
-        for library in libraries:
-            self[library.phase.name] = library
-
-    def __repr__(self):
-        return f"DiffractionLibrary<Phases:{list(self.keys())}>)"
 
 
 class DiffractionLibrary(dict):
